+++ CHANGELOG.md
@@ -38,6 +38,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Error when constrained voronoiDiagram called with `GROUP` PShape input.
 * Removing duplicate vertices during PShape->JTS conversion would remove every vertex (not just the duplicated ones).
 
+### Removed
+- `PGS_Contour.straightSkeletonSolub()` (didn't meet robustness standards)
+
 ## **1.1.3** *(2021-09-01)*
 ### Added
 - `diameter()` to `PGS_ShapePredicates`. Computes the diameter of a shape.